+++ examples/muse-layout-antd/tests/features/home/HeaderItem.test.js
@@ -3,6 +3,17 @@ import { shallow } from 'enzyme';
 import { HeaderItem } from '../../../src/features/home';
 
 it('renders node with correct class name', () => {
-  const renderedComponent = shallow(<HeaderItem />);
+  const renderedComponent = shallow(
+    <HeaderItem
+      meta={{
+        label: 'label',
+        link: 'link',
+        linkTarget: 'linkTarget',
+        onClick: jest.fn(),
+        icon: '',
+        className: '',
+      }}
+    />,
+  );
   expect(renderedComponent.find('.home-header-item').length).toBe(1);
 });