+++ packages/muse-craco-plugin/lib/handleMuseLocalPlugins.js
@@ -5,7 +5,7 @@ const { getLoader, loaderByName } = require('@craco/craco');
  * Allow webpack to support multiple local muse plugin projects.
  * So that it's possible to develop multiple projects together.
  *   1. Allow to find modules from plugin project's node_modules dir
- *   2. Allow to resolve muse lib package from any plugin project, for example @ebay/muse-react
+ *   2. Allow to resolve muse lib package from any plugin project, for example @ebay/muse-lib-react
  *   3. Tell babel to compile all project's source code
  * @param {*} webpackConfig
  * @returns
@@ -14,7 +14,7 @@ function handleMuseLocalPlugins(webpackConfig) {
   // Read local plugins from env variable
   const localPlugins = (process.env.MUSE_LOCAL_PLUGINS || '')
     .split(';')
-    .map((s) => s.trim())
+    .map(s => s.trim())
     .filter(Boolean);
   const babelInclude = [];
   const localPluginNames = [];
@@ -22,8 +22,8 @@ function handleMuseLocalPlugins(webpackConfig) {
   if (!Array.isArray(webpackConfig.entry)) webpackConfig.entry = [webpackConfig.entry];
 
   localPlugins
-    .map((p) => (path.isAbsolute(p) ? p : path.join(process.cwd(), p)))
-    .forEach((p) => {
+    .map(p => (path.isAbsolute(p) ? p : path.join(process.cwd(), p)))
+    .forEach(p => {
       // forced to src folder
       babelInclude.push(path.join(p, 'src'));
       const localPluginPkgJson = require(path.join(p, 'package.json'));
@@ -52,7 +52,9 @@ function handleMuseLocalPlugins(webpackConfig) {
   // Remove ModuleScopePlugin to support multiple src folders
   // TODO: maybe keep it by the similar mechanism expandPluginsScope at:
   // https://github.com/oklas/react-app-alias/blob/80c8a062b37df8411bc148dd42f485d014e96d3f/packages/react-app-alias/src/index.js#L26
-  const foundIndex = webpackConfig.resolve.plugins.map((x) => x.constructor.name).indexOf('ModuleScopePlugin');
+  const foundIndex = webpackConfig.resolve.plugins
+    .map(x => x.constructor.name)
+    .indexOf('ModuleScopePlugin');
   if (foundIndex >= 0) {
     webpackConfig.resolve.plugins.splice(foundIndex, 1);
   }