+++ packages/muse-craco-plugin/lib/jest/setupAfterEnv.js
@@ -1,0 +1,2 @@
+require('./__mocks__/museConfig.mock');
+require('./__mocks__/museEntries.mock');