+++ packages/muse-dev-utils/lib/index.js
@@ -2,5 +2,4 @@ module.exports = {
   museContext: require('./museContext'),
   utils: require('./utils'),
   setupMuseDevServer: require('./setupMuseDevServer'),
-  getDevMuseGlobal: require('./getDevMuseGlobal'),
 };