+++ packages/muse-plugin-react-scripts/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ebay/muse-plugin-react-scripts
 
+## 1.0.4
+
+### Patch Changes
+
+- Dev test.
+
+## 1.0.3
+
+### Patch Changes
+
+- Dev test.
+
 ## 1.0.2
 
 ### Patch Changes